+++ src/test/msixtest/api_bundlemanifestreader.cpp
@@ -2,120 +2,20 @@
 //  Copyright (C) 2019 Microsoft.  All rights reserved.
 //  See LICENSE file in the project root for full license information.
 // 
+//  Validates IAppxBundleManifestReader interface and interfaces related to the AppxBundleManifest.xml
 #include "catch.hpp"
 #include "msixtest_int.hpp"
 #include "FileHelpers.hpp"
 #include "UnbundleTestData.hpp"
-
 #include "macros.hpp"
 
-void InitializeBundleHelper(const std::string& package, IAppxBundleReader** bundleReader)
-{
-    *bundleReader = nullptr;
-
-    auto bundlePath = MsixTest::TestPath::GetInstance()->GetPath(MsixTest::TestPath::Directory::Unbundle) + "/" + package;
-    bundlePath = MsixTest::Directory::PathAsCurrentPlatform(bundlePath);
-
-    MsixTest::ComPtr<IAppxBundleFactory> bundleFactory;
-    MsixTest::ComPtr<IStream> inputStream;
-
-    REQUIRE_SUCCEEDED(CreateStreamOnFile(const_cast<char*>(bundlePath.c_str()), true, &inputStream));
-    REQUIRE_SUCCEEDED(CoCreateAppxBundleFactoryWithHeap(
-        MsixTest::Allocators::Allocate,
-        MsixTest::Allocators::Free,
-        MSIX_VALIDATION_OPTION::MSIX_VALIDATION_OPTION_SKIPSIGNATURE,
-        static_cast<MSIX_APPLICABILITY_OPTIONS>(MSIX_APPLICABILITY_OPTIONS::MSIX_APPLICABILITY_OPTION_SKIPPLATFORM |
-                                                MSIX_APPLICABILITY_OPTIONS::MSIX_APPLICABILITY_OPTION_SKIPLANGUAGE),
-        &bundleFactory));
-
-    REQUIRE_SUCCEEDED(bundleFactory->CreateBundleReader(inputStream.Get(), bundleReader));
-    REQUIRE_NOT_NULL(*bundleReader);
-    return;
-}
-
-// Validates a footprint files from a bundle
-TEST_CASE("Api_AppxBundleReader_FootprintFiles", "[api]")
-{
-    std::string bundle = "StoreSigned_Desktop_x86_x64_MoviesTV.appxbundle";
-    MsixTest::ComPtr<IAppxBundleReader> bundleReader;
-    InitializeBundleHelper(bundle, &bundleReader);
-
-    MsixTest::ComPtr<IAppxFile> appxBlockMap;
-    REQUIRE_SUCCEEDED(bundleReader->GetFootprintFile(MsixTest::Constants::Bundle::AppxBlockMap.first, &appxBlockMap));
-    MsixTest::Wrappers::Buffer<wchar_t> appxBlockMapName;
-    REQUIRE_SUCCEEDED(appxBlockMap->GetName(&appxBlockMapName));
-    REQUIRE(MsixTest::Constants::Bundle::AppxBlockMap.second == appxBlockMapName.ToString());
-
-    MsixTest::ComPtr<IAppxFile> appxBundleManifest;
-    REQUIRE_SUCCEEDED(bundleReader->GetFootprintFile(MsixTest::Constants::Bundle::AppxBundleManifest.first, &appxBundleManifest));
-    MsixTest::Wrappers::Buffer<wchar_t> appxBundleManifestName;
-    REQUIRE_SUCCEEDED(appxBundleManifest->GetName(&appxBundleManifestName));
-    auto expected = MsixTest::Constants::Bundle::AppxBundleManifest.second;
-    std::replace(expected.begin(), expected.end(), '/', '\\');
-    REQUIRE(expected == appxBundleManifestName.ToString());
-
-    MsixTest::ComPtr<IAppxFile> appxSignature;
-    REQUIRE_SUCCEEDED(bundleReader->GetFootprintFile(MsixTest::Constants::Bundle::AppxSignature.first, &appxSignature));
-    MsixTest::Wrappers::Buffer<wchar_t> appxSignatureName;
-    REQUIRE_SUCCEEDED(appxSignature->GetName(&appxSignatureName));
-    REQUIRE(MsixTest::Constants::Bundle::AppxSignature.second == appxSignatureName.ToString());
-}
-
-// Validates list of payload packages in the bundle
-TEST_CASE("Api_AppxBundleReader_PayloadPackages", "[api]")
-{
-    std::string bundle = "StoreSigned_Desktop_x86_x64_MoviesTV.appxbundle";
-    MsixTest::ComPtr<IAppxBundleReader> bundleReader;
-    InitializeBundleHelper(bundle, &bundleReader);
-
-    auto expectedPackages = MsixTest::Unbundle::GetExpectedPackages();
-
-    MsixTest::ComPtr<IAppxBundleReaderUtf8> bundleReaderUtf8;
-    REQUIRE_SUCCEEDED(bundleReader->QueryInterface(UuidOfImpl<IAppxBundleReaderUtf8>::iid, reinterpret_cast<void**>(&bundleReaderUtf8)));
-
-    MsixTest::ComPtr<IAppxFilesEnumerator> packages;
-    REQUIRE_SUCCEEDED(bundleReader->GetPayloadPackages(&packages));
-
-    BOOL hasCurrent = FALSE;
-    REQUIRE_SUCCEEDED(packages->GetHasCurrent(&hasCurrent));
-    int numOfPackages = 0;
-    while (hasCurrent)
-    {
-        MsixTest::ComPtr<IAppxFile> package;
-        REQUIRE_SUCCEEDED(packages->GetCurrent(&package));
-
-        auto expectedPackage = expectedPackages[numOfPackages];
-
-        MsixTest::Wrappers::Buffer<wchar_t> packageName;
-        REQUIRE_SUCCEEDED(package->GetName(&packageName));
-        REQUIRE(expectedPackage.name == packageName.ToString());
-
-        MsixTest::Wrappers::Buffer<char> packageNameUtf8;
-        MsixTest::ComPtr<IAppxFileUtf8> packageUtf8;
-        REQUIRE_SUCCEEDED(package->QueryInterface(UuidOfImpl<IAppxFileUtf8>::iid, reinterpret_cast<void**>(&packageUtf8)));
-        REQUIRE_SUCCEEDED(packageUtf8->GetName(&packageNameUtf8));
-        REQUIRE(expectedPackage.name == packageNameUtf8.ToString());
-
-        MsixTest::ComPtr<IAppxFile> package2;
-        REQUIRE_SUCCEEDED(bundleReader->GetPayloadPackage(packageName.Get(), &package2));
-        REQUIRE_ARE_SAME(package.Get(), package2.Get());
-
-        MsixTest::ComPtr<IAppxFile> package3;
-        REQUIRE_SUCCEEDED(bundleReaderUtf8->GetPayloadPackage(packageNameUtf8.Get(), &package3));
-        REQUIRE_ARE_SAME(package.Get(), package3.Get());
-
-        REQUIRE_SUCCEEDED(packages->MoveNext(&hasCurrent));
-        numOfPackages++;
-    }
-    REQUIRE(expectedPackages.size() == numOfPackages);
-}
 
 // Validates IAppxBundleManifestReader::GetStream
 TEST_CASE("Api_AppxBundleManifestReader_Stream", "[api]")
 {
     std::string bundle = "StoreSigned_Desktop_x86_x64_MoviesTV.appxbundle";
     MsixTest::ComPtr<IAppxBundleReader> bundleReader;
-    InitializeBundleHelper(bundle, &bundleReader);
+    MsixTest::InitializeBundleReader(bundle, &bundleReader);
     MsixTest::ComPtr<IAppxBundleManifestReader> bundleManifestReader;
     REQUIRE_SUCCEEDED(bundleReader->GetManifest(&bundleManifestReader));
 
@@ -129,7 +29,7 @@ TEST_CASE("Api_AppxBundleManifestReader_PackageId", "[api]")
 {
     std::string bundle = "StoreSigned_Desktop_x86_x64_MoviesTV.appxbundle";
     MsixTest::ComPtr<IAppxBundleReader> bundleReader;
-    InitializeBundleHelper(bundle, &bundleReader);
+    MsixTest::InitializeBundleReader(bundle, &bundleReader);
     MsixTest::ComPtr<IAppxBundleManifestReader> bundleManifestReader;
     REQUIRE_SUCCEEDED(bundleReader->GetManifest(&bundleManifestReader));
 
@@ -209,7 +109,7 @@ TEST_CASE("Api_AppxBundleManifestReader_PackageInfo", "[api]")
 {
     std::string bundle = "StoreSigned_Desktop_x86_x64_MoviesTV.appxbundle";
     MsixTest::ComPtr<IAppxBundleReader> bundleReader;
-    InitializeBundleHelper(bundle, &bundleReader);
+    MsixTest::InitializeBundleReader(bundle, &bundleReader);
     MsixTest::ComPtr<IAppxBundleManifestReader> bundleManifestReader;
     REQUIRE_SUCCEEDED(bundleReader->GetManifest(&bundleManifestReader));
